Move Code::Add() functions to their own file
