+++ src/generate/code.cpp
@@ -10,8 +10,10 @@
     Notes:
 
     The Eol() function will automatically append tabs if m_indent is greater than 0. That means you should *not* append tabs
-   using += '/t', and you should be very cautious about using += '\n' instead of Eol() since the Eol() call will
-   automatically append tabs if needed.
+    using += '/t', and you should be very cautious about using += '\n' instead of Eol() since the Eol() call will
+    automatically append tabs if needed.
+
+    code_add.cpp contains the Code::Add...() functions
 
 */
 
@@ -279,48 +281,9 @@ static const view_map s_short_rust_map
 {
 };
 
-constexpr auto set_perl_constants = frozen::make_set<std::string_view>({
-
-    "wxALL",
-    "wxLEFT",
-    "wxRIGHT",
-    "wxTOP",
-    "wxBOTTOM",
-
-    "wxEXPAND",
-    "wxSHAPED",
-    "wxFIXED_MINSIZE",
-    "wxRESERVE_SPACE_EVEN_IF_HIDDEN",
-
-    "wxALIGN_CENTER_HORIZONTAL",
-    "wxALIGN_CENTER_VERTICAL",
-    "wxALIGN_LEFT",
-    "wxALIGN_RIGHT",
-    "wxALIGN_TOP",
-    "wxALIGN_BOTTOM",
-    "wxALIGN_CENTER",
-
-    "wxITEM_CHECK",
-    "wxITEM_DROPDOWN",
-    "wxITEM_NORMAL",
-    "wxITEM_RADIO",
-
-    "wxNullBitmap",
-    "wxID_ANY",
-
-    "wxVERTICAL",
-    "wxHORIZONTAL",
-    "wxBOTH",
-
-    "wxWINDOW_VARIANT_LARGE",
-    "wxWINDOW_VARIANT_SMALL",
-    "wxWINDOW_VARIANT_MINI",
-
-});
-
 // clang-format on
 
-std::string_view GetLanguagePrefix(tt_string_view candidate, GenLang language)
+std::string_view Code::GetLanguagePrefix(tt_string_view candidate, GenLang language)
 {
     const view_map* prefix_list;
     const view_map* global_list;
@@ -637,49 +600,6 @@ void Code::CloseFontBrace()
     }
 }
 
-Code& Code::AddAuto()
-{
-    if (is_local_var())
-    {
-        if (is_cpp())
-        {
-            *this += "auto* ";
-        }
-        else if (is_perl())
-        {
-            *this += "my $";
-        }
-        else if (is_python())
-        {
-            return *this;  // no modifier for local variables in Python
-        }
-        else if (is_ruby())
-        {
-            return *this;  // no modifier for local variables in Ruby
-        }
-        else if (is_rust())
-        {
-            *this += "let ";
-        }
-#if GENERATE_NEW_LANG_CODE
-
-        else if (is_fortran())
-        {
-            *this += "type(";
-        }
-        else if (is_haskell())
-        {
-            *this += "let ";
-        }
-        else if (is_lua())
-        {
-            *this += "local ";
-        }
-#endif  // GENERATE_NEW_LANG_CODE
-    }
-    return *this;
-}
-
 void Code::InsertLineBreak(size_t cur_pos)
 {
     ASSERT(cur_pos > 1 && cur_pos <= size());
@@ -730,204 +650,12 @@ Code& Code::as_string(PropName prop_name)
     return Add(m_node->as_string(prop_name));
 }
 
-Code& Code::AddType(tt_string_view text)
-{
-    if (is_cpp() || text.size() < 3)
-    {
-        CheckLineLength(text.size());
-        *this += text;
-    }
-    else if (is_ruby())
-    {
-        auto new_text = ConvertToUpperSnakeCase(text.substr(2));
-        CheckLineLength(sizeof("Wx::") + new_text.size());
-        *this << "Wx::" << new_text;
-    }
-    else
-    {
-        CheckLineLength(m_language_wxPrefix.size() + text.size() - 2);
-        *this << m_language_wxPrefix << text.substr(2);
-    }
-
-    return *this;
-}
-
-Code& Code::AddConstant(tt_string_view text)
-{
-    if (is_cpp() || is_perl())
-    {
-        CheckLineLength(text.size());
-        *this += text;
-        return *this;
-    }
-    return Add(text);
-}
-
-Code& Code::Add(tt_string_view text)
-{
-    bool old_linebreak = m_auto_break;
-    // Ruby changes the prefix to "Wx::", and Python changes it to "wx."
-    // C++, Perl, and Rust use the constant unmodified.
-    //
-    // "wx" is the shortest string that could be changed -- no single letter will ever be changed by
-    // this function.
-    if (is_cpp() || is_perl() || is_rust() || text.size() < (sizeof("wx") - 1))
-    {
-        CheckLineLength(text.size());
-        *this += text;
-    }
-    else
-    {
-        if (is_ruby())
-        {
-            if (text.size())
-            {
-                // Ruby doesn't like breaking the parenthesis for a function call onto the next line,
-                // or the .new function
-                if (text.front() == '.' || text.front() == '(')
-                {
-                    old_linebreak = m_auto_break;
-                    m_auto_break = false;
-                }
-            }
-            if (text == "wxEmptyString")
-            {
-                // wxRuby prefers ('') for an empty string instead of the expected Wx::empty_string
-                *this += "('')";
-                return *this;
-            }
-            else if (text == "wxDefaultCoord")
-            {
-                *this += "Wx::DEFAULT_COORD";
-                return *this;
-            }
-            else if (text == "wxDefaultSize")
-            {
-                *this += "Wx::DEFAULT_SIZE";
-                return *this;
-            }
-            else if (text == "wxDefaultPosition")
-            {
-                *this += "Wx::DEFAULT_POSITION";
-                return *this;
-            }
-            else if (text == "wxNullBitmap")
-            {
-                *this += "Wx::NULL_BITMAP";
-                return *this;
-            }
-            else if (text == "wxNullAnimation")
-            {
-                *this += "Wx::NULL_ANIMATION";
-                return *this;
-            }
-        }
-
-        if (text.find('|') != tt::npos)
-        {
-            bool initial_combined_value_set = false;
-            tt_view_vector multistr(text, "|", tt::TRIM::both);
-            for (auto& iter: multistr)
-            {
-                if (iter.empty())
-                    continue;
-                if (initial_combined_value_set)
-                    *this += '|';
-                if (iter.is_sameprefix("wx") && !is_cpp())
-                {
-#if 0
-                    if (is_perl() && (HasPerlMapConstant(text) || set_perl_constants.contains(text)))
-                    {
-                        CheckLineLength(text.size());
-                        *this += text;
-                        initial_combined_value_set = true;
-                        continue;
-                    }
-#endif
-                    if (std::string_view language_prefix = GetLanguagePrefix(text, m_language); language_prefix.size())
-                    {
-                        // Some languages will have a module added after their standard prefix.
-                        CheckLineLength(language_prefix.size() + iter.size() - 2);
-                        *this << language_prefix << iter.substr(2);
-                    }
-                    else
-                    {
-                        // If there was no sub-language module added (e.g., wx.aui. for
-                        // Python), then use the default language prefix.
-                        CheckLineLength(m_language_wxPrefix.size() + iter.size() - 2);
-                        *this << m_language_wxPrefix << iter.substr(2);
-                    }
-                }
-                else
-                {
-                    CheckLineLength(iter.size());
-                    *this += iter;
-                }
-                initial_combined_value_set = true;
-            }
-        }
-        else if (text.is_sameprefix("wx") && !is_cpp())
-        {
-            if (is_perl())
-            {
-                if (HasPerlMapConstant(text) || set_perl_constants.contains(text))
-                {
-                    CheckLineLength(text.size());
-                    *this += text;
-                    return *this;
-                }
-                else if (text == "wxEmptyString")
-                {
-                    *this << "\"\"";
-                    return *this;
-                }
-
-                if (std::string_view language_prefix = GetLanguagePrefix(text, m_language); language_prefix.size())
-                {
-                    CheckLineLength(language_prefix.size() + text.size() - 2);
-                    *this << language_prefix << text.substr(2);
-                }
-                else
-                {
-                    CheckLineLength(m_language_wxPrefix.size() + text.size() - 2);
-                    *this << m_language_wxPrefix << text.substr(2);
-                }
-            }
-            else if (std::string_view language_prefix = GetLanguagePrefix(text, m_language); language_prefix.size())
-            {
-                CheckLineLength(language_prefix.size() + text.size() - 2);
-                *this << language_prefix << text.substr(2);
-            }
-            else
-            {
-                CheckLineLength(m_language_wxPrefix.size() + text.size() - 2);
-                *this << m_language_wxPrefix << text.substr(2);
-            }
-        }
-        else
-        {
-            CheckLineLength(text.size());
-            *this += text;
-        }
-    }
-
-    // In case linebreak was shut off
-    m_auto_break = old_linebreak;
-
-    return *this;
-}
-
 Code& Code::TrueFalseIf(GenEnum::PropName prop_name)
 {
     if (m_node->as_bool(prop_name))
         return True();
     else
         return False();
-}
-
-Code& Code::AddConstant(GenEnum::PropName prop_name, tt_string_view short_name)
-{
-    return Add(m_node->as_constant(prop_name, short_name));
 }
 
 Code& Code::Function(tt_string_view text, bool add_operator)
@@ -3063,31 +2791,6 @@ bool Code::is_ScalingEnabled(GenEnum::PropName prop_name, int enable_dpi_scaling
     return true;
 }
 
-Code& Code::AddComment(std::string_view comment, bool force)
-{
-    if (!UserPrefs.is_AddComments() && !force)
-        return *this;
-    Eol(eol_if_needed);
-    switch (m_language)
-    {
-        case GEN_LANG_CPLUSPLUS:
-            *this << "// " << comment;
-            break;
-        case GEN_LANG_PYTHON:
-        case GEN_LANG_RUBY:
-        case GEN_LANG_PERL:
-            *this << "# " << comment;
-            break;
-        default:
-            *this << "# " << comment;
-            break;
-    }
-
-    Eol(eol_if_needed);
-
-    return *this;
-}
-
 Code& Code::BeginConditional()
 {
     if (is_cpp() || is_perl())
@@ -3098,70 +2801,6 @@ Code& Code::BeginConditional()
     {
         *this << "if ";
     }
-    return *this;
-}
-
-Code& Code::AddConditionalAnd()
-{
-    if (is_cpp() || is_ruby() || is_perl() || is_rust())
-    {
-        *this << " && ";
-    }
-    else if (is_python())
-    {
-        *this << " and ";
-    }
-#if GENERATE_NEW_LANG_CODE
-    else if (is_fortran())
-    {
-        *this << " .AND. ";
-    }
-    else if (is_haskell())
-    {
-        *this << " && ";
-    }
-    else if (is_lua())
-    {
-        *this << " and ";
-    }
-#endif
-
-    else
-    {
-        MSG_WARNING("unknown language");
-    }
-
-    return *this;
-}
-Code& Code::AddConditionalOr()
-{
-    if (is_cpp() || is_ruby() || is_perl() || is_rust())
-    {
-        *this << " || ";
-    }
-    else if (is_python())
-    {
-        *this << " or ";
-    }
-#if GENERATE_NEW_LANG_CODE
-    else if (is_fortran())
-    {
-        *this << " .OR. ";
-    }
-    else if (is_haskell())
-    {
-        *this << " || ";
-    }
-    else if (is_lua())
-    {
-        *this << " or ";
-    }
-#endif
-    else
-    {
-        MSG_WARNING("unknown language");
-    }
-
     return *this;
 }
 
@@ -3223,41 +2862,4 @@ Code& Code::False()
     }
 
     return *this;
-}
-
-void Code::AddPublicRubyMembers()
-{
-    ASSERT(is_ruby());
-    std::set<tt_string> public_members;
-    auto FindPublicMembers = [&](Node* node, auto&& FindPublicMembers) -> void
-    {
-        if (node->hasProp(prop_var_name) && node->as_string(prop_class_access) == "public:")
-        {
-            public_members.insert(tt_string(":") << node->getNodeName(get_language()));
-        }
-        if (node->getChildCount())
-        {
-            for (const auto& child: node->getChildNodePtrs())
-            {
-                FindPublicMembers(child.get(), FindPublicMembers);
-            }
-        }
-    };
-
-    FindPublicMembers(m_node, FindPublicMembers);
-
-    if (public_members.size())
-    {
-        Indent(1);
-        Tab() << "attr_accessor ";
-        for (auto iter = public_members.begin(); iter != public_members.end(); ++iter)
-        {
-            if (iter != public_members.begin())
-                *this << ", ";
-            *this << *iter;
-            CheckLineLength();
-        }
-        ResetIndent();
-        Eol();
-    }
 }
